Fix contest and choice standardization
Updating the values of the <select> elements at render-time is
problematic, because re-renders will reset values as the background
polling takes place.

This uses React component state to hold the state of the form, and
modifies the <select> elements to be controlled components to ensure
that the selected value from the dropdown matches what is actually in
the Page component's state.

Initialization (and, well, the code) is still not ideal, but doesn't
violate React norms as wildly as abusing a public object member to hold
state or updating state at render time (!)